Updated topic_stats.py: Altered the university & uni by region functions for both primary & secondary routines. Now those functions are by university and region separately.
The data being displayed wasn't weighted originally which gave a false impression of the popularity of topics. This has been rectified by displaying the proportional share of each topic as a percentage of total modules per given query rather than the numerical count.

Changed some labels and added more comments to clarify the purpose of certain code blocks.

Modified the module codes to have the university acronym suffixed due to issues with multiple table joins which caused modules with the same codes at different universities to be associated together.